Created Checklist class. Moved all classes to separate files. Debug argument now has default value
Created a checklist class with check(config) method, which does the check of the whole checlist against config which was produced base on the user's config file
Moved OR and OptConfig classes to a separate file
Moved Outputter class to its own file
Renamed Config class to UserConfig and moved it to a separate file
